# Performance: remove useless header merge from Response