--- a/lib/rack/utils.rb
+++ b/lib/rack/utils.rb
@@ -410,15 +410,9 @@ module Rack
 
     # A case-insensitive Hash that preserves the original case of a
     # header when set.
-    class HeaderHash < Hash
-      def self.build(hash)
-        HeaderHash === hash ? hash.dup : new(hash)
-      end
-
-      def self.new(hash = {})
-        HeaderHash === hash ? hash : super(hash)
-      end
-
+    #
+    # @api private
+    class HeaderHash < Hash # :nodoc:
       def initialize(hash = {})
         super()
         @names = {}