## Change

There is no need to instantiate and merge headers, as `HeaderHash` already sets the given values from `header`.

By looking at the code of `HeaderHash#merge` we see that it duplicates itself:

```ruby
      def merge(other)
        hash = dup
        hash.merge! other
      end
```

Each time we instantiate a `Rack::Response` we instantiate a `HeaderHash` twice: the first time because of `new`, and the second time because of `merge`.

## Req/s benchmark

By running this benchmark on my machine, I can see a performance gain of +350req/s.

**Gemfile**

```ruby
# frozen_string_literal: true

source "https://rubygems.org"

gem "rack" #, git: "https://github.com/jodosha/rack.git", branch: "performance/remove-useless-header-merge"
gem "puma"
```

**config.ru**

```ruby
# frozen_string_literal: true

require "bundler/setup"

class MyAction
  def call(*)
    res = Rack::Response.new
    res.write "OK"
    res
  end
end

run MyAction.new
```

Run the app with:

```shell
$ bundle exec puma -e production -t 16:16 config.ru
```

**Before:**

```shell
$ wrk -t 2 http://localhost:9292/
Running 10s test @ http://localhost:9292/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.96ms  520.27us  17.09ms   86.12%
    Req/Sec     5.33k   317.62     5.67k    91.50%
  106150 requests in 10.00s, 4.05MB read
Requests/sec:  10614.81
Transfer/sec:    414.65KB
```

**After:**

```shell
$ wrk -t 2 http://localhost:9292/
Running 10s test @ http://localhost:9292/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.93ms  525.01us  16.51ms   87.49%
    Req/Sec     5.53k   358.19     5.85k    92.00%
  109938 requests in 10.00s, 4.19MB read
Requests/sec:  10993.58
Transfer/sec:    429.45KB
```