@tenderlove 2019-01-15
I have a few thoughts about this:

1. I think the subclass `HeaderHash` should behave as close to the super as possible, meaning `.new` should dup
2. I really don't like that `HeaderHash` is a subclass of `Hash`.  I'd prefer it delegates to `Hash`
3. `HeaderHash` should be `:nodoc:` and private.

I can't find any usage of HeaderHash on GitHub code search that *isn't* internal to Rack, so I think we're pretty safe to change this class.  I don't expect anyone to be calling `HeaderHash.new`.

@jodosha is it possible to just delete our current implementation of `self.new` on `HeaderHash`?