@jodosha 2019-01-23
@tenderlove Please have a look at https://github.com/rack/rack/pull/1325/commits/d671a0c3379bc05f17e3f670d943c6243c333df0, this commit:

  1. It removes `Utils::HeaderHash.new` and `.build` (previously introduced in this PR)
  1. It changes an existing test to assert that all the given headers (both `Hash` and `Utils::HeaderHash`) will be duped.
  1. It marks `Utils::HeaderHash` as private API and adds a `:nodoc:` directive for RDoc.

---

Regarding `Utils::HeaderHash` inheriting from `Hash`, I agree. Do you think there is room for this improvement in this PR, or it's better to wait for this change? Thanks.