@tenderlove 2019-02-03
> Regarding `Utils::HeaderHash` inheriting from `Hash`, I agree. Do you think there is room for this improvement in this PR, or it's better to wait for this change? Thanks.

I guess we could consider this PR a "breaking change" already, so I'm fine with removing the inheritance in this PR.

We haven't decided on the version for master, whether it's 2.1 or 3.0.  Maybe we should merge this in and call it 3.0?  @rafaelfranca @matthewd @eileencodes @jeremy any opinions?  As far as performance goes, this HeaderHash thing has been a thorn in our side for a while.  I'd be really happy to ship this.