@tenderlove 2019-02-03
To be clear, I think we could merge this PR as-is and ship it as 2.1.  If we change inheritance to delegation we'll have to go to 3.0 as the API surface area will change drastically (there's no way we can change to delegation and have even close to 100% api compatibility with `Hash`)