@jodosha 2019-02-20
@tenderlove 2.1 vs 3.0: it really depends on the general plans for Rack.

If you're planning other major changes for 3.0, it's worth to modify `Utils::HeaderHash` and use delegation, otherwise it can wait for now. Thanks.