@jodosha 2019-03-06
@tenderlove I realized that asking for Rack plans is out of the scope for this PR. So here's my proposal: if the current implementation looks good, please merge it.

I'll take care after to create another PR that introduces the breaking changes. So it's ready if you folks decide one day to go for 3.0.