@tenderlove 2019-04-02
@jodosha sorry to take so long to respond.  I'll merge this.

> I'll take care after to create another PR that introduces the breaking changes. So it's ready if you folks decide one day to go for 3.0.

This sounds great.  If we decide master should be 3.0 we can merge that PR.